gen_testing_report/requirements.txt
atlassian-python-api==3.21.0
certifi==2021.10.8
charset-normalizer==2.0.12
chinesecalendar==1.7.2
Deprecated==1.2.13
idna==3.3
oauthlib==3.2.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
urllib3==1.26.9
wrapt==1.14.0
